+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More rewrite rules for PEq, PNeg, missing eqByte call, and distributivity for And
 - UNSAT cache is now in `Solvers.hs` and is therefore shared across all threads.
   Hence, it is now active even during branch queries.
+- Allow changing of the prefix from "prove" via --prefix in `test` mode
+- More complete simplification during interpretation
 
 ## Fixed
 - We now try to simplify expressions fully before trying to cast them to a concrete value
@@ -110,6 +112,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   increases which hurt concrete execution performance due to their linear cost.
 - The concrete MCOPY implementation has been optimized to avoid freezing the whole
   EVM memory.
+- We no longer accept `check` as a prefix for test cases by default.
 
 ## [0.54.2] - 2024-12-12